+++ CHANGELOG.md
@@ -2,9 +2,10 @@
   * [ExUnit] Use ANSI escape codes in CLI output
   * [ExUnit] Include suite run time on CLI results
   * [IEx] Support `pwd` and `cd` helpers
-  * [Macro] `Macro.escape` now supports `escape_unquote` as an option
   * [Kernel] Better error reporting for invalid bitstring generators
   * [Kernel] Improve meta-programming by allowing `unquote` on `def/2`, `defp/2`, `defmacro/2` and `defmacrop/2`
+  * [Macro] `Macro.escape` now supports `escape_unquote` as an option
+  * [Mix] Support `--cover` on mix test and `test_coverage` on Mixfiles
 
 * bug fix
   * [Binary] inspect no longer escapes standalone hash `#`